requirements.txt
streamlit
faiss-cpu
numpy
pickle-mixin
sentence-transformers
transformers
pypdf
